requirements.txt
pytest
flake8
black
fastapi[all]
pydantic
pylint
psycopg2
sqlalchemy
asyncpg
anyio
asyncmock
asyncio
trio
pytest-asyncio
pytest-cov
